move package endpoint tests to HttpServiceIntegrationTest
